Invert symbols and invert crossings
Crossing key renamed to `crossing convention`, key added to replace
symbols by their inverses.  This should cover the majority of variants!

Also, a `nowarning` option for the original braid package to suppress
the warning about it being old.

Documentation of all of the above, plus an example showing the various
lengths.